refactor: split rules injector processor facade
